backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx~=0.27.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
selenium
webdriver-manager
sqlalchemy==2.0.23
lyricsgenius==3.0.1

yt-dlp
pytoniq>=0.1.38
